only works with properly formed URLs
RAWMSG - A block of text that may or may not contain a URL (change this as necessary to fit your application).
URLREGEX - Returns an array with the position and length of a URL found in MESSAGE (if no URL is present both values are 0). 
URLSTRING - harvests the URL from MESSAGE using the position and length values of URLREGEX.
HYPERLINK - an HTML hyperlink containing the harvested URL contained in URLSTRING.
HTMLMSG - The final product, if the original (RAWMSG) contained a URL this will contain an HTML hyperlink which leads to said URL.

The snippet simply uses Regular Expressions to determine if a block of plain text contains a URL and replaces the URL with an HTML hyperlink. It's not much, but it is a base which can be built upon.